Create `/patterns` and move UI patterns there
