+++ src/components/app/Features.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Stack from 'stack-styled';
 import VisuallyHidden from '../core/VisuallyHidden';
-import Feature from './Feature';
+import Feature from '../patterns/Feature';
 import SvgFeatureDuck from './images/SvgFeatureDuck';
 import SvgFeatureFireplace from './images/SvgFeatureFireplace';
 import SvgFeatureMegaphone from './images/SvgFeatureMegaphone';
