--- src/components/app/Features.md
@@ -1,3 +1,0 @@
-```jsx harmony
-<Features />
-```
