+++ src/components/patterns/DogCard/index.js
@@ -1,0 +1,1 @@
+export { default } from './DogCard';
