+++ src/components/patterns/Feature/Feature.md
@@ -1,5 +1,5 @@
 ```jsx harmony
-import SvgFeatureDuck from './images/SvgFeatureDuck';
+import SvgFeatureDuck from '../../app/images/SvgFeatureDuck';
 <Feature imageComponent={SvgFeatureDuck} title="Rubberducking">
   Expain your problem to a dog, not to a rubber duck.
 </Feature>;
