+++ src/components/patterns/Feature/index.js
@@ -1,0 +1,1 @@
+export { default } from './Feature';
