+++ src/components/patterns/Footer/Footer.md
